Add files via upload
Practice for a Request Mapping as  a parent mapping and child mapping
here is the scenario: create a new sillyController.java to use the same mapping("/showForm"), but this mapping already in the HelloWorldController.java, so when you run the app, it has ambiguous mapping problem ,  it did not know which to process. 

==== 
Solve: add  a @RequestMapping outside the class as a parent mapping("/hello") in HelloWorldController.java, and change the link jsp helloworld.jsp <a href="/hello/showForm"></a> it will use the HelloWorldController.java mapping one.